+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2023-04-26
 ### Added
 - Standard Noteable open source patterns
   - Contributing / Code of Conduct files
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `WebsocketManager` backend uses vanilla `logging` instead of `structlog`, remove need for `structlog` dependency once `managed-service-fixtures` also drops it
 - `JupyterBackend` introduce a short sleep in its poll loop while investigating 100% CPU usage
 - `JupyterBackend` zmq polling changed fairly significantly to avoid missing messages while reconnecting socket after a max message size disconnect
+- Try to include tracebacks during exception logs for inbound / outbound / process workers
 
 ## [0.2.2] - 2022-07-28
 ### Changed